+++ examples/HF135/inverse1.py
@@ -26,48 +26,48 @@ epsilon = 1e-7
 # Wetting of an HF135 membrane, Van Genuchten model
 # Data from Buser (PhD thesis, 2016)
 # http://hdl.handle.net/1773/38064
-S_range = (0.0473, 0.945)
+theta_range = (0.0473, 0.945)
 k = 5.50e-13  # m**2
 alpha = 0.2555  # 1/m
 n = 2.3521
-Si = 0.102755  # Computed from P0
+theta_i = 0.102755  # Computed from P0
 
-Sb = S_range[1] - epsilon
+theta_b = theta_range[1] - epsilon
 
-D_analytical = van_genuchten(n=n, alpha=alpha, k=k, theta_range=S_range)
+D_analytical = van_genuchten(n=n, alpha=alpha, k=k, theta_range=theta_range)
 
-analytical = solve(D=D_analytical, i=Si, b=Sb)
+analytical = solve(D=D_analytical, i=theta_i, b=theta_b)
 
 o = np.linspace(analytical.o[0], analytical.o[-1], 2000)
 
-S = analytical(o=o)
+theta = analytical(o=o)
 
-D_inverse = inverse(o=o, samples=S)
+D_inverse = inverse(o=o, samples=theta)
 
 
 fig = plt.figure()
 fig.canvas.set_window_title("Diffusivity plot")
 
 plt.title("Diffusivity function")
-plt.plot(S, D_analytical(S), label="Analytical")
-plt.plot(S, D_inverse(S), label="Obtained with inverse()")
-plt.xlabel("saturation [-]")
+plt.plot(theta, D_analytical(theta), label="Analytical")
+plt.plot(theta, D_inverse(theta), label="Obtained with inverse()")
+plt.xlabel("water content [-]")
 plt.ylabel("diffusivity [{}**2/{}]".format(validation.r_unit, validation.t_unit))
 plt.yscale('log')
 plt.grid(which='both')
 plt.legend()
 
 
-inverse = solve(D=D_inverse, i=Si, b=Sb, verbose=2)
+inverse = solve(D=D_inverse, i=theta_i, b=theta_b, verbose=2)
 
 fig = plt.figure()
-fig.canvas.set_window_title("Saturation plot")
+fig.canvas.set_window_title("Water content plot")
 
-plt.title("Saturation in terms of o")
+plt.title("Water content in terms of o")
 plt.plot(o, analytical(o=o), label="Using analytical diffusivities")
 plt.plot(o, inverse(o=o), label="Using diffusivities obtained with inverse()")
 plt.xlabel("o [{}/{}**0.5]".format(validation.r_unit, validation.t_unit))
-plt.ylabel("saturation [-]")
+plt.ylabel("water content [-]")
 plt.grid(which='both')
 plt.legend()