+++ examples/HF135/inverse2.py
@@ -17,22 +17,22 @@ from fronts import solve, inverse, o
 
 import validation
 
-D_inverse = inverse(o=o(validation.r, validation.t)[::5], samples=validation.S[::5])
+D_inverse = inverse(o=o(validation.r, validation.t)[::5], samples=validation.theta[::5])
 # Using only a fifth of the points so that it does not run too slow
 
-S = solve(D=D_inverse, i=validation.S[-1], b=validation.S[0], itol=5e-3, verbose=2)
+theta = solve(D=D_inverse, i=validation.theta[-1], b=validation.theta[0], itol=5e-3, verbose=2)
 
 
 fig = plt.figure()
-fig.canvas.set_window_title("Saturation plot")
+fig.canvas.set_window_title("Water content plot")
 
-plt.title("Saturation field at t={} {}".format(validation.t, validation.t_unit))
-plt.plot(validation.r, validation.S,
+plt.title("Water content field at t={} {}".format(validation.t, validation.t_unit))
+plt.plot(validation.r, validation.theta,
          label="Original ({})".format(validation.name))
-plt.plot(validation.r, S(validation.r, validation.t),
+plt.plot(validation.r, theta(validation.r, validation.t),
          label="Reconstructed with inverse() and solve()")
 plt.xlabel("r [{}]".format(validation.r_unit))
-plt.ylabel("saturation [-]")
+plt.ylabel("water content [-]")
 plt.grid(which='both')
 plt.legend()