+++ examples/HF135/radial.py
@@ -21,20 +21,19 @@ epsilon = 1e-7
 # Wetting of an HF135 membrane, Van Genuchten model
 # Data from Buser (PhD thesis, 2016)
 # http://hdl.handle.net/1773/38064
-S_range = (0.0473, 0.945)
+theta_range = (0.0473, 0.945)
 k = 5.50e-13  # m**2
 alpha = 0.2555  # 1/m
 n = 2.3521
-Si = 0.102755  # Computed from P0
+theta_i = 0.102755  # Computed from P0
 h = 1.60e-4  # m -- thickness
-eps = 0.76  # porosity
 
 flowrate = 1e-9  # m**3/s
 
-D = van_genuchten(n=n, alpha=alpha, k=k, theta_range=S_range)
+D = van_genuchten(n=n, alpha=alpha, k=k, theta_range=theta_range)
 
 
-S = solve_flowrate(D=D, i=Si, Qb=flowrate/eps, radial='cylindrical', height=h,
+theta = solve_flowrate(D=D, i=theta_i, Qb=flowrate, radial='cylindrical', height=h,
 	               verbose=2)
 
 
@@ -42,13 +41,13 @@ r = np.linspace(0, 5e-2, 200)  # m
 t = (60, 120)  # s
 
 fig = plt.figure()
-fig.canvas.set_window_title("Saturation plot")
+fig.canvas.set_window_title("Water content plot")
 
-plt.title("Saturation fields")
-plt.plot(r, S(r,t[0]), label="t={} {}".format(t[0], t_unit))
-plt.plot(r, S(r,t[1]), label="t={} {}".format(t[1], t_unit))
+plt.title("Water content fields")
+plt.plot(r, theta(r,t[0]), label="t={} {}".format(t[0], t_unit))
+plt.plot(r, theta(r,t[1]), label="t={} {}".format(t[1], t_unit))
 plt.xlabel("position [{}]".format(r_unit))
-plt.ylabel("saturation [-]")
+plt.ylabel("water content [-]")
 plt.grid(which='both')
 plt.legend()
 
@@ -56,10 +55,10 @@ fig = plt.figure()
 fig.canvas.set_window_title("Velocity plot")
 
 plt.title("Velocity fields")
-plt.plot(r, S.flux(r,t[0]), label="t={} {}".format(t[0], t_unit))
-plt.plot(r, S.flux(r,t[1]), label="t={} {}".format(t[1], t_unit))
+plt.plot(r, theta.flux(r,t[0]), label="t={} {}".format(t[0], t_unit))
+plt.plot(r, theta.flux(r,t[1]), label="t={} {}".format(t[1], t_unit))
 plt.xlabel("position [{}]".format(r_unit))
-plt.ylabel("true velocity [{}/{}]".format(r_unit, t_unit))
+plt.ylabel("Darcy velocity [{}/{}]".format(r_unit, t_unit))
 plt.grid(which='both')
 plt.legend()
 
@@ -67,9 +66,9 @@ fig = plt.figure()
 fig.canvas.set_window_title("Flow rate plot")
 
 plt.title("Flow rate fields")
-plt.plot(r, S.flux(r,t[0]) * eps * (2*pi*r) * h,
+plt.plot(r, theta.flux(r,t[0]) * (2*pi*r) * h,
          label="t={} {}".format(t[0], t_unit))
-plt.plot(r, S.flux(r,t[1]) * eps * (2*pi*r) * h,
+plt.plot(r, theta.flux(r,t[1]) * (2*pi*r) * h,
          label="t={} {}".format(t[1], t_unit))
 plt.xlabel("position [{}]".format(r_unit))
 plt.ylabel("flow rate [{}**3/{}]".format(r_unit, t_unit))