+++ examples/HF135/solve.py
@@ -21,29 +21,29 @@ epsilon = 1e-7
 # Wetting of an HF135 membrane, Van Genuchten model
 # Data from Buser (PhD thesis, 2016)
 # http://hdl.handle.net/1773/38064
-S_range = (0.0473, 0.945)
+theta_range = (0.0473, 0.945)
 k = 5.50e-13  # m**2
 alpha = 0.2555  # 1/m
 n = 2.3521
-Si = 0.102755  # Computed from P0
+theta_i = 0.102755  # Computed from P0
 
-Sb = S_range[1] - epsilon
+theta_b = theta_range[1] - epsilon
 
-D = van_genuchten(n=n, alpha=alpha, k=k, theta_range=S_range)
+D = van_genuchten(n=n, alpha=alpha, k=k, theta_range=theta_range)
 
 
-S = solve(D=D, i=Si, b=Sb, verbose=2)
+theta = solve(D=D, i=theta_i, b=theta_b, verbose=2)
 
 
 fig = plt.figure()
-fig.canvas.set_window_title("Saturation plot")
+fig.canvas.set_window_title("Water content plot")
 
-plt.title("Saturation field at t={} {}".format(validation.t, validation.t_unit))
-plt.plot(validation.r, S(validation.r,validation.t),
+plt.title("Water content field at t={} {}".format(validation.t, validation.t_unit))
+plt.plot(validation.r, theta(validation.r,validation.t),
          color='steelblue', label="Fronts")
-plt.plot(validation.r, validation.S, color='sandybrown', label=validation.name)
+plt.plot(validation.r, validation.theta, color='sandybrown', label=validation.name)
 plt.xlabel("position [{}]".format(validation.r_unit))
-plt.ylabel("saturation [-]")
+plt.ylabel("water content [-]")
 plt.grid(which='both')
 plt.legend()
 
@@ -52,12 +52,12 @@ fig = plt.figure()
 fig.canvas.set_window_title("Velocity plot")
 
 plt.title("Velocity field at t={} {}".format(validation.t, validation.t_unit))
-plt.plot(validation.r, S.flux(validation.r,validation.t),
+plt.plot(validation.r, theta.flux(validation.r,validation.t),
          color='steelblue', label="Fronts")
 plt.plot(validation.r, validation.velocity,
          color='sandybrown', label=validation.name)
 plt.xlabel("position [{}]".format(validation.r_unit))
-plt.ylabel("true velocity [{}/{}]".format(validation.r_unit, validation.t_unit))
+plt.ylabel("Darcy velocity [{}/{}]".format(validation.r_unit, validation.t_unit))
 plt.grid(which='both')
 plt.legend()